Fixed tests and moved useDropdown logic into AbstractDropdown. - RSC-674
